Fix for bug#1744 - MPIConnection is now able to send messages larger than 2GG. If a message larger than 2GB is sent it will be broken up into chunks of 2GB and sent in sequence.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@2447 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Base/mwbase/mwcommon/trunk/mwcommon/MWConnection.h
@@ -30,6 +30,7 @@
 #ifndef ASKAP_MWCOMMON_MWCONNECTION_H
 #define ASKAP_MWCOMMON_MWCONNECTION_H
 
+#include <unistd.h>
 #include <boost/shared_ptr.hpp>
 
 //# Forward Declarations
@@ -107,11 +108,11 @@ namespace askap { namespace mwbase {
 
     /// Receive the given amount of data in the buffer.
     /// and wait until data has been received into buf.
-    virtual void receive (void* buf, unsigned size) = 0;
+    virtual void receive (void* buf, size_t size) = 0;
 
     /// Send the fixed sized data to the connected MWConnection
     /// and wait until the data has been sent.
-    virtual void send (const void* buf, unsigned size) = 0;
+    virtual void send (const void* buf, size_t size) = 0;
   };
 
 }} /// end namespaces
